refactor(api): split configsync_apply.go by concern
configsync_apply.go held the import orchestration, the fences and
settings validation, the model-intent apply and the per-entity upserts
in one file. Two seams move out, same package:

- configsync_apply_models.go: modelStateApplyTimeout, modelIntentWriter,
  applyModelIntent, applyDisabledModels, applyEnabledModels,
  writeUnappliedModelRefs
- configsync_apply_upserts.go: applyFailoverGroups,
  providerTypeForImport, upsertProviders, applyUsers, usernameToID,
  upsertVirtualKeys, groupApplyResult, upsertFailoverGroups,
  providerNameToID

Pure moves with their doc comments, imports pruned. configsync_apply.go
goes from 1161 to 496 lines and leaves the size allowlist.
